Rename azure_private_dns azureprivatedns
